requirements.txt
alembic==1.9.0
Babel==2.11.0
blinker==1.5
certifi==2022.12.7
chardet==3.0.4
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
dominate==2.7.0
elastic-transport==8.4.0
elasticsearch==8.5.3
email-validator==1.3.0
Flask==2.2.2
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.0
Flask-Moment==1.0.5
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
googletrans==3.0.0
greenlet==2.0.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
importlib-metadata==5.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
packaging==22.0
pipdeptree==2.3.3
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.7
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==1.4.45
urllib3==1.26.13
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.11.0

# requirements for Heroku
#psycopg2==2.7.3.1
#gunicorn==19.7.1